Add CM support

+++ test/e2e/storage/utils/create.go
@@ -290,6 +290,7 @@ var factories = map[What]ItemFactory{
 	{"ServiceAccount"}:     &serviceAccountFactory{},
 	{"StatefulSet"}:        &statefulSetFactory{},
 	{"StorageClass"}:       &storageClassFactory{},
+	{"ConfigMap"}:          &cmFactory{},
 }
 
 // PatchName makes the name of some item unique by appending the
@@ -569,6 +570,27 @@ func (*storageClassFactory) Create(f *framework.Framework, i interface{}) (func(
 	client := f.ClientSet.StorageV1().StorageClasses()
 	if _, err := client.Create(context.TODO(), item, metav1.CreateOptions{}); err != nil {
 		return nil, errors.Wrap(err, "create StorageClass")
+	}
+	return func() error {
+		return client.Delete(context.TODO(), item.GetName(), &metav1.DeleteOptions{})
+	}, nil
+}
+
+type cmFactory struct{}
+
+func (f *cmFactory) New() runtime.Object {
+	return &v1.ConfigMap{}
+}
+
+func (*cmFactory) Create(f *framework.Framework, i interface{}) (func() error, error) {
+	item, ok := i.(*v1.ConfigMap)
+	if !ok {
+		return nil, errorItemNotSupported
+	}
+
+	client := f.ClientSet.CoreV1().ConfigMaps(f.Namespace.GetName())
+	if _, err := client.Create(context.TODO(), item, metav1.CreateOptions{}); err != nil {
+		return nil, errors.Wrap(err, "create ConfigMap")
 	}
 	return func() error {
 		return client.Delete(context.TODO(), item.GetName(), &metav1.DeleteOptions{})
